isolate images tests from files, and disable them for now
