+++ tests/api-tests/fields/images.crud.test.disabled.ts
@@ -6,7 +6,7 @@ import fetch from 'node-fetch';
 // @ts-ignore
 import Upload from 'graphql-upload/Upload.js';
 import mime from 'mime';
-import { file, text, image } from '@keystone-6/core/fields';
+import { text, image } from '@keystone-6/core/fields';
 import { list } from '@keystone-6/core';
 import { KeystoneConfig, StorageConfig } from '@keystone-6/core/types';
 import { setupTestRunner } from '@keystone-6/api-tests/test-runner';
@@ -93,148 +93,6 @@ const checkFile = async (
     return fs.existsSync(path.join(config.folder, filename));
   }
 };
-
-describe('File - Crud special tests', () => {
-  const filename = 'keystone.jpeg';
-  const fileHash = createHash('sha1')
-    .update(fs.readFileSync(path.resolve(fieldPath, 'image/test-files', filename)))
-    .digest('hex');
-
-  const createItem = (context: any) =>
-    context.query.Test.createOne({
-      data: { secretFile: prepareFile(filename, 'image') },
-      query: `
-        id
-        secretFile {
-          filename
-          __typename
-          filesize
-          url
-        }
-    `,
-    });
-
-  for (let matrixValue of testMatrix) {
-    const getConfig = (): StorageConfig => ({
-      ...(matrixValue === 's3'
-        ? { ...s3DefaultStorage, preserve: false, type: 'file' }
-        : {
-            kind: 'local',
-            type: 'file',
-            storagePath: fs.mkdtempSync(path.join(os.tmpdir(), 'file-local-test')),
-            serverRoute: { path: '/files' },
-            generateUrl: (path: string) => `http://localhost:3000/files${path}`,
-          }),
-    });
-
-    const fields = { secretFile: file({ storage: 'test_file' }) };
-
-    describe(matrixValue, () => {
-      describe('Create - upload', () => {
-        const config = getConfig();
-        const hashConfig: { matrixValue: 'local'; folder: string } | { matrixValue: 's3' } =
-          config.kind === 'local'
-            ? { matrixValue: 'local', folder: `${config.storagePath}/`! }
-            : { matrixValue: config.kind };
-        test(
-          'Upload values should match expected corrected',
-          getRunner({ storage: { test_file: { ...config } }, fields })(async ({ context }) => {
-            const data = await createItem(context);
-            expect(data).not.toBe(null);
-
-            expect(data.secretFile).toEqual({
-              /*
-                The url and filename here include a hash, and currently what we are doing
-                is just checking the url is modified correctly - that said, this sucks
-                as a test
-                */
-              url:
-                matrixValue === 's3'
-                  ? expect.stringContaining(`/${data.secretFile.filename}`)
-                  : `http://localhost:3000/files/${data.secretFile.filename}`,
-              filename: data.secretFile.filename,
-              __typename: 'FileFieldOutput',
-              filesize: 3250,
-            });
-            // check file exists at location
-            expect(fileHash).toEqual(await getFileHash(data.secretFile.filename, hashConfig));
-          })
-        );
-      });
-
-      describe('After Operation Hook', () => {
-        const config = getConfig();
-        const hashConfig =
-          config.kind === 'local'
-            ? { matrixValue: config.kind, folder: `${config.storagePath}/`! }
-            : { matrixValue: config.kind };
-        test(
-          'with preserve: true',
-          getRunner({ storage: { test_file: { ...config, preserve: true } }, fields })(
-            async ({ context }) => {
-              const {
-                id,
-                secretFile: { filename },
-              } = await createItem(context);
-
-              expect(await checkFile(filename, hashConfig)).toBeTruthy();
-
-              const {
-                secretFile: { filename: filename2 },
-              } = await context.query.Test.updateOne({
-                where: { id },
-                data: { secretFile: prepareFile('thinkmill.jpg', 'file') },
-                query: `secretFile { filename }`,
-              });
-
-              expect(await checkFile(filename, hashConfig)).toBeTruthy();
-              expect(await checkFile(filename2, hashConfig)).toBeTruthy();
-
-              await context.query.Test.deleteOne({ where: { id } });
-
-              expect(await checkFile(filename, hashConfig)).toBeTruthy();
-              // TODO test that just nulling the field doesn't delete it
-            }
-          )
-        );
-        test(
-          'with preserve: false',
-          getRunner({
-            storage: { test_file: { ...config, preserve: false } },
-            fields,
-          })(async ({ context }) => {
-            const {
-              id,
-              secretFile: { filename },
-            } = await createItem(context);
-
-            expect(await checkFile(filename, hashConfig)).toBeTruthy();
-
-            const {
-              secretFile: { filename: filename2 },
-            } = await context.query.Test.updateOne({
-              where: { id },
-              data: { secretFile: prepareFile('thinkmill.jpg', 'file') },
-              query: `
-                secretFile {
-                  filename
-                }`,
-            });
-
-            expect(await checkFile(filename2, hashConfig)).toBeTruthy();
-            expect(await checkFile(filename, hashConfig)).toBeFalsy();
-
-            await context.query.Test.deleteOne({ where: { id } });
-
-            expect(await checkFile(filename2, hashConfig)).toBeFalsy();
-
-            // TODO test that just nulling the field removes the file
-          })
-        );
-      });
-    });
-  }
-});
 
 describe('Image - Crud special tests', () => {
   const createItem = async (context: any, filename: string) =>
